requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-multipart==0.0.6
pycryptodome==3.19.0
phe==1.5.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
google-cloud-storage==2.13.0
google-cloud-secret-manager==2.16.4

# 测试依赖
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1
aiofiles==23.2.1
pytest-mock==3.12.0
async-asgi-testclient==1.4.11 